upgrade ruff, fix stdlib-module-shadowing (A005)
upgrading started enforcing this A005 rules, which was cranky
about our Python packages named "json" and "types". It was easy
enough to just rename those packages, and everything still seems
to work.
