+++ .pre-commit-config.yaml
@@ -6,7 +6,7 @@ repos:
 
   - repo: https://github.com/astral-sh/ruff-pre-commit
     # Ruff version.
-    rev: v0.9.4
+    rev: v0.9.6
     hooks:
       - id: ruff
         args: [ --fix ]
