+++ src/ims/model/jsons/_accessentry.py
