+++ src/ims/model/jsons/_entry.py
